Implement first Democrit GSP version.
This implements a very basic GSP for Democrit, tracking executed trades
by seller name and the seller-provided ID.  The game state is simply the
ever-growing set of executed trades.  Moves are of the following form:

  {"g": {"dem": {"t": "seller id"}}}

For now, the GSP only processes confirmed moves and allows querying
of the total game state for testing purposes.  Later on, it will also
track pending moves and allow a query of "get current state for name/id",
returning one of "unknown", "pending", "confirmed".  This will be the
main interface used by Democrit.