Implement basic MUC XMPP client.
This starts implementation of the actual Democrit library.  As a first
step, we define a basic class for our XMPP connection, including joining
a MUC room.

This commit adds the basic MucClient class, with support for connecting,
joining, leaving and disconnecting (but not yet any other of the
functionality we will need later on).