Track full JIDs for nicknames.
Add functionality to the MucClient to keep track of the real, full JIDs
of room participants per nickname.  We use that to map nicknames to
full JIDs when processing messages, so that we can use the full JIDs
(which e.g. may be XID-authenticated by the XMPP server) to identify
peers rather than their self-chosen nicknames.