StanzaExtensions for our protocol buffers.
This implements a gloox StanzaExtension that will hold serialised
OrdersOfAccount protocol buffers, namely for the broadcast messages
sent by each user with their own orders to the MUC channel.

The implementation itself is generic as a template, so that it can be
used for other stanza types with other protocol buffer payloads as well
in the future.