Locking of own orders.
Add functionality to MyOrders that allows temporarily locking (and later
unlocking) orders.  Locked orders will be marked as such in the own orders
returned by RPC, and will not be broadcast as available -- but they can be
restored easily again by unlocking.

This will be used for orders that are currently being taken by some
counterparty but before the counterparty has shared signatures, i.e.
as a temporary measure to prevent race conditions while being able to
quickly restore the orders if the counterparty is not going through.