Update docs:  Taker signs first.
Update the description of the protocol in the README file.  Instead of the
seller, the taker (seller or buyer) should sign "first".  This ensures
that the maker, who has to at least temporarily remove the order from
the orderbook, is in charge of the final signature and thus can't as easily
be subjected to a DoS attack.