Update README to use btxid.
Update the protocol description in the README file:  Instead of using
a taker-provided custom ID to track the trade, use the bare hash (btxid).
This is more robust and simpler.

Also add descriptions of when to lock/unlock outputs in the wallet
during the trade execution to avoid accidentally spending them.