Define Trade wrapper class.
This adds the new Trade and TradeManager classes, which take care of
interpreting and processing the underlying TradeState protocol buffer
data.

For now we can just create new trades based on an existing offer that
is taken (either by us or from us), and extract the "public info" of
a trade in form of a Trade proto.  In the future, the classes will also
handle the processing and updates to the state done via messages
with the counterparty.