Default for MyOrders::UpdateOrders.
Instead of making MyOrders::UpdateOrders pure virtual, provide a default
implementation (that does nothing).  This makes subclassing it easier
e.g. in tests, but more importantly, also fixes a potential issue when
an instance is destructed:  When a MyOrders instance is destructed, the
refresher thread (which calls into UpdateOrders) is still running until
the destructor is finished.  This means that it may try to call UpdateOrders
after the subclass is already destructed, which previously led to a call
on a pure method and aborted.