Infrastructure for trade ProcessingMessage's.
This adds the basic infrastructure for Trade's to get updated with
incoming messages from the counterparty, and to build up their own
reply messages for the counterparty.

For now, it supports processing and building a message that tells
a maker about the taker's intend to start a trade, but nothing else
is implemented yet.