Exchange seller data for trades.
This extends the processing logic for trades, so that it supports
the seller sending their addresses (and recording for themselves
the name output being used).

For this, we need an RPC connection to the Xaya wallet; for now with
getnewaddress and name_show.  We use a mock implementation for unit
testing.