Verification logic for buyer.
This implements the verification logic ensuring, for a buyer, that a trade
is safe to sign.  In particular, it checks that the seller can actually send
the assets, in a way that guarantees that the trade cannot go wrong if the
right outpoint is used as name input in the trade transaction.