Logic for constructing the trade tx.
This adds Trade::ConstructTransaction, which is the logic the buyer
executes to build up the unsigned trade transaction on their end based
on their own wallet and the seller data.

For now, this method is just prepared as well as tested basically.  In the
future, it will be integrated in the actual trade-processing flow together
with the buyer and seller checks using TradeChecker.