Construct, share and sign PSBTs.
This adds the message handling / processing logic to actually construct
PSBTs as the buyer, share PSBTs with the counterparty, sign them and
update the state accordingly.

All the steps sharing PSBTs as per the trade protocol are implemented, but
for now the final step where the maker combines the signatures and actually
broadcasts the transaction is still missing.  That will be a follow-up
commit.