Implement signature checking for the buyer.
This adds TradeChecker::CheckForBuyerSignature:  The method compares the
PSBTs before and after the buyer signed, and verifies that exactly all
but one input have been signed.

While the buyer constructs the transaction themselves and is thus a bit
better protected against cheating than the seller, the seller might still
impersonate a name that's in the buyer's wallet.  This check ensures that
in this case, the buyer would not proceed with the trade (signing everything
and effectively trading against themselves).