Time out trades if not finalised.
When trades are "initiated" but do not turn "pending" (i.e. the processing
is finalised and we start waiting on network confirmations), time them out
to "abandoned" after a relatively short period.

This processing should only take a couple of seconds at most, since it just
involves the off-chain signing and no confirmations.  It is also easy and
free to abandon in this case, as we have not yet shared our signatures and
can just stop processing the trade.