Detect double-spend failures for trades.
When updating a trade, not only detect success but also failure if one
of the inputs has been double spent (i.e. is no longer available in the
UTXO set).  When this happens first, we remember the current block height;
when it is still conflicted some blocks after, we mark it as "confirmed
failed" so it can be archived.