Update myorders for finalised trades.
If a trade is finalised (succeeds / abandoned / failed) and we are the
maker, update MyOrders accordingly:  If the trade failed, we unlock and
thus "restore" the original order so it is available again.  If the trade
succeeded, we try to restore a partial order if it has only been filled
partially and the remaining quantity is still enough for the specified
min_units.