Add TradeManager to Daemon instance.
This extends the Daemon class to hold (internally) a TradeManager instance.
For this, we need to pass through a Xaya and g/dem GSP RPC URL in the
various parts.

With the TradeManager, the Daemon will be able to return a list of active
trades and handle requests to take an order.  For both of them, we also have
new public RPC methods.

For now, the TradeManager is not yet integrated with XMPP, so it won't
actually send messages to the trade counterparty nor process received
messages; that will come in the next step.