Return full GSP state data for checktrade.
In the Democrit GSP's checktrade method, return the full, usual
GSP state data including the GSP's block height, block hash and syncing
state.  The trade's state is returned in the "data" field.

This makes the method fit in with other GSPs we have built, and also
allows the trade daemon to know the GSP's block height from the call
(which it needs for keeping proper state of conflicted trades).