Instantly set state's account.
The account name is something that a lot of code expects to be set
always in the state proto.  With this change, we make sure it is set
right in the constructor of the State instance; this avoids potential
race conditions with other threads (e.g. the MyOrders refresher) starting
up already while we have not yet set the account name in the Daemon
constructor body.