Do not start Daemon in constructor.
Previously, the Daemon class' constructor immediately established
the connection to the XMPP server and started the reconnecting thread.

With this change, we split this out into a new function Connect, which
has to be called explicitly after the constructor.  This gives users more
flexibility, and allows e.g. in the future to configure the instance a bit
more before "starting" it.