Expect value_encoding to be UTF-8.
Xaya Core will use UTF-8 as its default value encoding starting
with 1.6.0, so it makes sense to require/expect this.  With UTF-8,
all values that are valid according to consensus rules can be
encoded, so that external applications using the RPC interface get
all the data and can do their own processing.

In the case of Democrit, the value string is compared against a
locally-generated "expected" value string anyway, so there is no way
an attacker can try to trick a parser with weird data.  In concrete
terms, this means that values encoded as actual UTF-8 won't be accepted
by Democrit anyway, since the local JSON encoding step of the expected
move data will not generate such data.